Bug fix: non-english dictionaries and decoder
The CHAR_SPACE and WORD_SPACE characters were not being used in the non-English dictionaries so they wouldn't translate any more.
The decoder was also not using them so was unable to distinguish between char and word space lengths.

In fixing this, the constants have finally been moved into a separate file and included everywhere they are needed.